linkedin-scraper/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
pandas==2.1.3
python-dotenv==1.0.0
